refactor(PolicyDetails): RHICOMPL-1627 common EditRulesButtonToolbarItem
adds the Edit Rules button also to the rules view without the OS tabs
